Was a bug were tanks speed did not affect movement speed of division.
My theory: When the radio equipment is added to the batallion the unit does not know which of the equipments set the speed.

So I set "transport =" for the tanks.

Please test this:

Start game.   ->  do the instant research command ->  research heavy armor. -> research tank radio techs.   -> put heavy armor in motorized template.