Lots of Romania changes, mostly to the political tree (it's bigger and more historical), but also some smaller changes to Romania here and there. Production updated so that Romania starts with uniforms and other necessities in production, national focus factories should not be placed in Bessarabia anymore, some party names changed around, Romania starts democratic (but becomes non-aligned and then fascist), 1937 election event added, some cosmetic stuff where Romania doesn't become the "Legionary State" when the Iron Guard has been purged by Antonescu and instead remains the "Kingdom of Romania".
Oh, and King Michael's Coup actually does what it did in history now: It makes it possible for Romania to switch sides if the war with the Soviets is not going well (that is, if Bessarabia is in Soviet hands after 1941). Maybe not the most elegant code, but I think it does the job?