USA changes. Two kinds of changes here: First, focus AI changes. This should have the US AI pick smarter focuses. It will work hard to get rid of the Great Depression effects, but it will also try to enforce the embargo on Japan ASAP. If tension gets high enough, it will rush to get Giant Wakes and similar important focuses. When Barbarossa happens, the USA will work to get them lend lease aid ASAP (probably December 1941 or January 1942).
The second set of changes affects focus times. In general, trying to make US focuses shorter, removing a week here and a week there. Currently the USA has a lot of focuses that they ought to do, but really don't have time for. I don't want the AI to follow a strictly optimal multiplayer build either, so the AI will take focuses that are less useful but historically accurate. Shorter focuses help make that viable.