UK cosmetic focus changes. Purely cosmetic changes to the positioning of several focuses. First, the "Chain Home" and "Air Defense" focuses were not connected to the focuses above them. Instead of "fixing" that and possibly nerfing the UK, I just moved the focuses a bit so that things make sense again.
Then there's the "Singapore Strategy" and "Defence in Depth" focuses, both which require "Britannia Rules the Wave" but were above that focus. I fixed that by moving the focuses a bit.

Lastly, the "Royal Navy" focus (and the ones below that) was a bit lower than it had to be, so I raised that one a bit.