German-Japanese cooperation focuses fixed.
Japan and Germany share a lot of near-identical focuses. I've made it so that these are now mutually exclusive with each other. If Germany shares synthetic resource technology with Japan, Japan will bypass the focus that has them ask for synthetic resource technology with Germany. I've also updated the localisation for Japan, since most of these focuses did not have any on the Japanese versions.

This also means that if Germany creates the Tripartite Pact, Japan will not have to create the Tripartite Pact, and vice versa.