"Extra Supplies" ability fix/change.
Previously, this ability added "Supply Grace" for 7 days. This effectively meant that your troops do not need supplies for 7 days (-100% supply use for 7 days), which allows for encircled troops to regain their org and avoid lack of supply maluses. If exploited hard enough, with the right generals and country bonuses, it could allow for troops to completely ignore supply for months.

I changed this to lower supply consumption by 50% for 14 days. Encircled troops will now remain out of supply, but the ability is still useful for troops that are connected to supplies in some way. If you're trying to put a lot of troops in Egypt, for example, this will still work pretty well.