USA changes.
Some focuses were removed (War Plan White, War Plan Red) since they were effectively useless. Plan Dog Memo was deactivated, because it has uninteresting effects and its description does not fit with where it currently is. Might move it around later. Pacific Basing was also deactivated, because it is pretty boring.

Some bypasses were added to certain focuses (Selective Service System is bypassed if you already have that conscription level or above).

The navy part of the focus tree was reworked a bit, adding a submarine focus as well as the historical torpedo issues that the USA had (thanks to BICE for the inspiration).