Soviet fix. If you do a more lenient army purge, the purge effects wouldn't gradually disappear because the wrong purge level was set (or the wrong event was set, depending on your perspective).
I increased the starting purge level to fix it. I tried to change the event instead at first, but that lead to issues with the purge ending too early, and I didn't want to mess around with that.

In conclusion, if the Soviets do the more lenient army purge (where both Tukhachevsky and Rokossovsky survive) the effects will start a bit harsher than before but they will gradually disappear now.