Mobilisation laws fix.
A lot of code that doesn't do anything except jumble up the tooltips was removed. This should make the demobilisation decisions more understandable, especially for France. I could have tried to fix the code so that it actually works, but that would be pretty major change that might impact countries (such as France) in unknown ways.