Various minor fixes
Hides some Hungarian focuses that do nothing.
Makes a lot of Romanian focuses shorter
Removes an unnecessary "mutually exclusive" in the UK focus tree which keeps confusing people
Disables SS support companies, because the duplicate support company issue keeps getting exploited by BezMemow
Fixes the German-Soviet peace so that they actually make peace with everyone in the Axis, not just Germany.
Increases local supplies in the Alexandria by 10, because it's a bit silly how low it is and all the various hoops that the UK has to jump to improve supply.