Minor changes to Japan.
Disabling the Manchukuo recognition focus for now, it can probably be implemented into the Manchukuo focus tree later.

Also added a tooltip for Ascend the Throne, removed incorrect tooltip from Chinese Interior, and made the Yamato battleships start in a more complete state to compensate for me nerfing them in a previous update. Also fixed the Yamato ship tooltip.

Made the "Expand the Military Power" focus shorter, since it doesn't really do a whole lot and to make it a bit easier for Japan to finish vital focuses in time.