Non-La Resistance Spanish Civil War changes
This adds a couple extra divisions for Nationalist Spain for people who don't have the DLC. Apparently there's an issue with Republican Spain winning the war without the DLC and without player involvement. I haven't tested if this actually fixes anything, but it's a quick and dirty improvement that should help a little bit at least.