Soviet fixes
Removes a duplicate light bomber for Soviets (it was literally a duplicate tech, unlocked the exact same equipment).

Also fixes some Soviet focuses so that they work better if a specific state is lost or already full with factories.

Also adds a Swedish embargo on the Soviets when the Soviets finish the Winter War focus.

Also fixes an Albanian port location.