UK focus path tweaks, and a minor USA change
The UK focus tree is a bit of a mess when it comes to the focus paths, this cleans it up. Focus paths are now much clearer, there are fewer overlapping paths, fewer paths that jump through focuses, fewer "surprise, you actually also need to do this focus!" situations, optional focuses are now actually optional (coughOverseas Militarycough), and it's now possible for UK players to access to the Pacific theatre focuses.

I didn't change what any of the UK focuses do, just tweaked some positions and changed the required focuses.

For the USA, I lowered the requirement for the Cash and Carry focus because it's often annoyingly difficult for the USA to unlock that focus.