Soviet resistance buff
Increases the base resistance value in Soviet territories with the NKVD idea from Beria's Partisans. Currently Soviet resistance is pitiful, worse than both the French and Polish, this should at least make it a bit more troublesome.

This should make the Russians resist harder than the French, but not as hard as the Poles.