Sino-Japanese overhaul
Slow zones are back (for PvP, PvAI and AIvAI)
Three levels of slow zones are now implemented. The harshest one is now reserved only for the interior parts of China, and two new less harsh variants for the coastal areas and areas closer to Japan.
Japan strategy improved for both Sino-Japanese war and the Pacific War
A timer is added for the stalemate, so that it is clear when it will happen. This timer can be extended by escalating the conflict or taking key cities (Beijing, Shanghai, Guangzhou, Wuhan and Changsha).
Japanese chemical warfare is slightly buffed and now also applies to ComChi, but now also increases tension by 1.
China focus tree improved, fewer restrictions on which focuses you can pick.

Things which I will need to add:
Missions for China so that they can see which cities they need to defend (Beijing, Shanghai, Guangzhou, Wuhan, Changsha) and when the stalemate will happen.
An anti-acclimatisation idea so that the slow zones don't lose their potency in warm/cold climates (I have one ready, but it's not implemented yet. You can look inside the Japanese ideas file to have a look at it though).