Minor tweaks
Adds missions to China so that a China player is aware (more or less) of what's going on with the stalemate mechanics, when the stalemate will happen, and which cities to defend to make the stalemate happen sooner.

Added localisation for all of this.

AI won't shorten factory shifts if they are at war, because that hurts their production (which they presumably need while they are at war).