Tweak to demilitarisation decision and generic war propaganda decision
Demilitarisation got a massive buff in a previous update, this brings it back to where it was before.

The generic war propaganda decision is now far easier to access since the targeted war propaganda decisions are gone, and it can also be used to reach much higher war support levels... since the targeted war propaganda decisions are gone.