Sorry, but I just can't stand this focus
The focus makes no sense, the Montreux Convention focus already exists, I don't understand what the focus description is saying, there was no massive migration from Ankara to Istanbul in this period, Istanbul was already referred to as Istanbul in 1936, and it's just a boring focus that most people will skip past anyway. There is nothing redeeming about it.

If you want to add it back, add it near the bottom of the Ottoman Empire alt-history stuff as a "Restore Istanbul as Capital of the Ottoman Empire" focus, and then have the whole migration stuff where people move back to the new/old capital, and increase stability.