Syria tweaks
Italy gets an option to annex Syria if a) they have taken Palestine/Jordan and border Syria, b) they are not at war with Syria, c) none of their enemies control Baghdad. Basically, if Italy wins in Africa they can take Syria.
Free France can't do the coup in Syria if they are at war with whoever controls Syria.

Basically, this gives the Axis an option to get through Syria and it makes it harder for surprise swaps of territories to happen, hopefully.

Also tells the Italian AI to form Italian East Africa, because why not.