Changing base reluctance back from 30 to 20
This will allow players to ask for control of states from the AI, and to give states to the AI, and probably some other things. Basically, it gives the player more control of the AI.

Maybe it has some unintended consequences, but they're probably not worse than seeing Falangist Poland steal half of Russia.