Minor fixes
Education I increases the starting level of generals and admirals, but does not increase the starting stats. That means that it's more or less a straight nerf to recruiting generals. Added extra starting stats to that tech. Considering making it hidden to make the effect tooltip less obnoxious.

Communist China had too high conscription levels for no clear reason, so I lowered that one step. Did someone fix the mutiny bug (mutinies never going away) that this was causing?