Reducing battalion weight of, mostly, artillery
Weight = the number of convoys required to transport divisions
When divisions get moved across the sea, they have to be escorted. For every convoy, you need at least 3 screens to have 100% screening efficiency. If your positioning is low (which it will be against submarines), you'll need even more screens for 100% screening efficiency.

Basically, a single division with 6 weight (which is extremely low for an infantry division), would need at least 18 destroyers to be properly screened. This explains some of the difficulties people have with submarines when they move units around.

This change reduces the weight of artillery divisions by roughly half (they're one of the bigger culprits), and rounding down some weights because you can't see if something weighs 0.2 or 0.25 in the division designer.