China map overhaul (part 2) TEST THIS
Splits up the Beijing/Shandong supply zone into two.
Splits up the Kaifeng/Wuhan supply zone into two.

Do some testing with this, compare it to the Steam version, see if the supply numbers (for Japan, China, and China when their capital is moved to Chongqing) make sense and are not broken. I imagine some of the local supply numbers should be decreased