USA focus fixes...
Try to avoid breaking the focus pathing in the future. If you can't see which focus leads to what, you're not creating a focus tree. You're just creating a squiggly jiggly abomination with random lines everywhere.

Also, remember to add new extremely important focuses to the AI focus list, or else the AI will be broken.