Remove army corruption 4...
If we're going to remove the positive effects of army_corruption_4, it makes little sense to keep the negative parts that were introduced to compensate for the positives. Just remove the whole idea, return back to the original state of the army corruption setup, and call it a day.

Basically, don't half-ass the revert. Revert it entirely.