France focus tree additions
Adds three new focuses to France, based on player feedback. One focus in case France manages to successfully advance and hold in a German border state for the duration of the focus, another focus to ban communism after the Molotov-Ribbentrop Pact is signed (historical), and one focus that lets France buy a larger amount of fighters from the USA

Plus localisation for it all

Honestly not super happy with any of these focuses. The "Not So Phony War" focus potentially encourages France players to be too aggressive with Germany, the "Ban Communism" focus already kinda exists but only for the ahistorical democratic path, and the "Islands for Security" focus is semi-historical but afaik they never actually gave up the islands. The American planes did arrive though. But hey, it's more France content.