Minor fixes
Finland's Sales Tax focus can now be done when you have any amount of government debts, not just maximum government debt

Raj's Land Reforms will now get rid of the Low Risk of Famine, not just the normal risk of famine

Guangdong's capital moved so that there aren't two Nanning's on the map. (why is the city of Nanning in a different state from the state of Nanning? oh well)