Testing enabling the Shanghai focus for AI Japan
Tweaks the Japan focus order a bit so that Marco Polo happens at a slightly more historical timing (earlier)

Reduces the time of the Shanghai focus for the same reasons

Also reducing the time for (one of) the "Proletariat Reserves" focuses that Soviets have, increasing the time for the Soviet Polish People's Army focus to compensate.

Fixed an annoying typo that doesn't show up in game.