This change is dedicated to the brave Mujahideen of Ireland
Allies now need to control Paris and Normandy to annex Vichy France

I considered having it be Paris and some combination of other coastal states, but a lot of those states either have a significant amount of VPs not on the actual coast, or are not neighbouring Paris and would therefore not rule out paratrooper shenanigans entirely. Normandy is both next to Paris and has most of its VPs on the coast, so it's a fairly safe choice.