Bugfix for crash
Each ideology has a unique occupation law (Local Autonomy for democracies, for example)
When you switch ideology, that occupation law is no longer valid and you'll revert to the default law (No Garrison). No Garrison is not available because of the restrictions added to it, so the game freezes.

Issue only seems to apply when the unique occupation law is applied on a specific country, not on a state or as the default occupation law.

The No Garrison restriction doesn't work anyway, you can get around it by selecting "No Garrison" as your default occupation law. Since No Garrison isn't available, it'll revert to the default law... which is No Garrison.