Raj improvements
Raj's "Investments in the Indian Army" focus can now be done before the UK has done the Overseas Military focus, however the UK can't do the decisions tied to that focus until they have done the Overseas Military focus. Tooltip added to explain this. This means that the Raj won't be locked out of several of their focuses if the UK hasn't yet completed an optional focus.