Rocket artillery nerfs
The 1944 rocket artillery tech is a bit too strong, and can easily be researched in the first half of 1942 by both Germany and Soviets thanks to their research bonuses.

This removes the 1 year ahead of time bonuses for Soviets and Germany, and increases the normal research bonus from 50% to 75% to compensate slightly.

The 25% soft attack bonus from the 1944 tech is removed entirely, and instead integrated into the 1944 rocket artillery equipment. This means that your ancient 1940 rocket artillery won't suddenly become better than 1944 light artillery just because of that one tech, you actually have to build modern rocket artillery.

Also, increased the cost of 1944 rocket artillery. It's now the same cost as 1944 light artillery (which effectively means that the battalion is slightly more expensive because rocket artillery uses 2 more horses than normal artillery??). 1944 rocket artillery is essentially superior in every way to 1944 light artillery, which is fine, but it should at least not be cheaper too.