Soviet changes
Order 227 is a focus instead of a decision, the idea is permanent instead of temporary. The effect is reduced from 15% territorial pride attack/defence to 8% territorial pride attack/defence.
Other focus times reduced to compensate for the new focus.

Collectivist Propaganda effect changed. It no longer gives temporary territorial pride, which was used to make the Soviets stronger than they should be in 1941, and instead provides a small amount of weekly stability, war support, and communist support over a long period of time. The end result should be a weaker Soviet army in 1941, but a slightly stronger industry to compensate. The Collectivist Propaganda focus will also be more viable as an early focus.

Changed Soviet infantry template from decisions (like Winter Offensive) slightly. It's now 20 width instead of 22 width, to make it more compatible with a standard Soviet division template. To compensate for the lost equipment, some extra guns are spawned. Are SOV_winteroffense_north.txt and SOV_winteroffense_south.txt used anywhere? Unsure, but changed them anyway just to be sure.