This is for Blood's Sweden changes, should be more or less functional
Germany's Treaty with Sweden now has a third option where Sweden can express interest in aligning closer with the Axis (as opposed to accepting the treaty while staying neutral, or just plain rejecting the treaty)
If that choice is picked, it unlocks the Approach Germany focus for Sweden.

The Approach Germany focus is also unlocked if Sweden is already fascist. Blood can remove that if he wants to.